requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
alabaster==0.7.16
async-timeout==4.0.3
attrs==23.2.0
Babel==2.14.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==7.1.2
click-didyoumean==0.3.0
contourpy==1.2.0
cycler==0.12.1
docutils==0.20.1
filelock==3.13.1
fonttools==4.47.2
frozenlist==1.4.1
fsspec==2023.12.2
huggingface-hub==0.20.3
idna==3.6
imagesize==1.4.1
Jinja2==3.1.3
kiwisolver==1.4.5
lightning-utilities==0.10.1
MarkupSafe==2.1.4
matplotlib==3.8.2
monai==1.3.0
mpmath==1.3.0
multidict==6.0.5
networkx==3.2.1
numpy==1.26.3
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
opencv-python==4.9.0.80
p==1.4.0
packaging==23.2
pandas==2.2.0
pillow==10.2.0
psutil==5.9.8
py-cpuinfo==9.0.0
pydicom==2.4.4
Pygments==2.17.2
pynetdicom==2.0.2
pyparsing==3.1.1
python-dateutil==2.8.2
pytorch-lightning==2.2.0
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
safetensors==0.4.2
scipy==1.12.0
seaborn==0.13.1
six==1.16.0
snowballstemmer==2.2.0
Sphinx==7.2.6
sphinxcontrib-applehelp==1.0.8
sphinxcontrib-devhelp==1.0.6
sphinxcontrib-htmlhelp==2.0.5
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.7
sphinxcontrib-serializinghtml==1.1.10
sympy==1.12
thop==0.1.1.post2209072238
timm==0.9.12
tokenizers==0.15.1
torch==2.2.0
torchmetrics==1.3.0.post0
torchvision==0.17
tqdm==4.66.1
transformers==4.37.2
triton==2.2.0
typing_extensions==4.9.0
tzdata==2023.4
ultralytics==8.1.5
urllib3==2.1.0
yarl==1.9.4
